GeometryParticle_csにてtimerの値がおかしいバグ修正 CPU型のtimerの型が間違っていた

+++ Source/Shader/ParticleSystem.cpp
@@ -426,7 +426,7 @@ void ParticleSystem::Render()
 }
 
 void ParticleSystem::Set(
-	float timer,
+	int timer,
 	DirectX::XMFLOAT3 p,
 	DirectX::XMFLOAT3 v,
 	DirectX::XMFLOAT3 f,
